feat : 비활성화 상태로 추가
